fetch

+++ rust/common/mod.rs
@@ -16,8 +16,8 @@ pub type Result<T = ()> = std::result::Result<T, Error>;
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
 pub struct LineColumn {
-    pub line: usize,
-    pub column: usize,
+    pub line: u32,
+    pub column: u32,
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
